+++ tests/database/factories/MechanicFactory.php
@@ -39,7 +39,7 @@ class MechanicFactory extends Factory
     public function definition()
     {
         return [
-            'name' => $this->faker->name,
+            'name' => $this->faker->name(),
         ];
     }
 }