+++ tests/database/factories/RoleFactory.php
@@ -39,7 +39,7 @@ class RoleFactory extends Factory
     public function definition()
     {
         return [
-            'name' => $this->faker->colorName,
+            'name' => $this->faker->colorName(),
         ];
     }
 }